requirements.txt
mysql-connector==2.2.9
mysql-connector-python==8.0.16
requests==2.22.0